+++ Source/WebArd/Commands/Session.cpp
@@ -1,0 +1,17 @@
+
+#include "ESP8266.hpp"
+
+
+int sessionId(){
+    return esp8266.read() - 48;
+}
+
+
+void closeSession(int connection){
+
+    String command = "AT+CIPCLOSE=";
+    command += connection;
+    command += "\r\n";
+
+    sendData(command,1000);
+}
